optimuspy-wavefront_subpath/automation/requirements.txt
pyyaml
munch
requests
bayesian-optimization
scikit-learn
scikit-optimize
diskcache
tqdm
torch
optuna
hyperopt